requirements.txt
Django>=5.0
PyPDF2>=3.0
python-docx>=1.0
reportlab>=4.0
PyMuPDF>=1.22
Pillow>=10.0